fixed some level 4 warnings

+++ Libraries/RobsJuceModules/rapt/Spectral/HarmonicAnalyzer.cpp
@@ -314,7 +314,7 @@ void rsHarmonicAnalyzer<T>::convertTimeUnit(RAPT::rsSinusoidalModel<T>& mdl)
 {
   for(size_t hi = 0; hi < mdl.getNumPartials(); hi++)
     for(int di = 0; di < getNumDataPoints(); di++)
-      mdl.getDataRef(hi, di).time /= sampleRate;
+      mdl.getDataRef((int)hi, di).time /= sampleRate;
 }
 
 template<class T>
